+++ packages/package-json-lint/src/rules/index.js
@@ -1,6 +1,6 @@
 const requireThemeContextVersions = require('./require-theme-context-versions');
 const requireNoTerraBasePeerDependencyVersions = require('./require-no-terra-base-peer-dependency-versions');
-const requireNoUnnecessaryDependencies = require('./require-no-unnecessary-dependencies');
+const requireNoUnnecessaryDependencies = require('./require-no-unnecessary-dependency');
 const requireNoHardCodedDependencyVersions = require('./require-no-hard-coded-dependency-versions');
 
 module.exports = {