+++ packages/package-json-lint/src/rules/require-no-unnecessary-dependency.js
@@ -4,83 +4,90 @@ const devDependencySet = [
   '@babel/helpers',
   '@babel/plugin-proposal-object-rest-spread',
   '@babel/plugin-syntax-async-generators',
+  '@babel/plugin-transform-object-assign',
   '@babel/plugin-transform-regenerator',
   '@babel/plugin-transform-runtime',
   '@babel/preset-env',
   '@babel/preset-react',
   '@babel/register',
-  '@cerner/terra-application',
-  '@cerner/terra-dev-site',
-  '@octokit/core',
-  'bufferutil',
-  'commander',
-  'danger',
-  'enzyme',
-  'enzyme-adapter-react-16',
-  'enzyme-to-json',
-  'eslint',
-  'html-webpack-plugin',
-  'jest',
-  'lerna',
-  'link-parent-bin',
-  'memory-fs',
-  'postcss',
-  'react',
-  'react-dom',
-  'react-intl',
-  'stylelint',
-  'utf-8-validate',
-  'webpack',
-  'webpack-cli',
-  'webpack-dev-server',
-  '@cerner/terra-cli',
-  'strip-ansi',
-  'yargs',
-  'chai',
-  'lodash.groupby',
-  'lodash.map',
-  'lodash.mapkeys',
-  'lodash.mapvalues',
-  'mocha',
-  'nock',
-  'np',
-  'sinon',
-  '@babel/plugin-transform-object-assign',
   '@cerner/browserslist-config-terra',
   '@cerner/eslint-config-terra',
   '@cerner/jest-config-terra',
   '@cerner/stylelint-config-terra',
   '@cerner/terra-aggregate-translations',
+  '@cerner/terra-application',
+  '@cerner/terra-cli',
+  '@cerner/terra-dev-site',
   '@cerner/terra-functional-testing',
   '@cerner/terra-open-source-scripts',
   '@cerner/webpack-config-terra',
+  '@octokit/core',
+  'bufferutil',
+  'chai',
   'check-installed-dependencies',
+  'commander',
   'core-js',
+  'csvtojson',
+  'danger',
+  'enzyme',
+  'enzyme-adapter-react-16',
+  'enzyme-to-json',
+  'eslint',
   'gh-pages',
   'glob',
+  'html-webpack-plugin',
+  'jest',
+  'jsdom',
   'ky',
+  'lerna',
+  'link-parent-bin',
+  'lodash.groupby',
+  'lodash.map',
+  'lodash.mapkeys',
+  'lodash.mapvalues',
+  'memory-fs',
+  'mocha',
+  'nock',
+  'np',
+  'one-cerner-style-icons',
+  'postcss',
   'react-test-renderer',
   'regenerator-runtime',
+  'shelljs',
+  'sinon',
+  'strip-ansi',
+  'stylelint',
+  'svgo',
   'terra-application',
+  'terra-button',
   'terra-disclosure-manager',
   'terra-enzyme-intl',
+  'utf-8-validate',
+  'webpack',
+  'webpack-cli',
+  'webpack-dev-server',
   'webpack-merge',
-  'csvtojson',
-  'jsdom',
-  'one-cerner-style-icons',
-  'shelljs',
-  'svgo',
-  'terra-button',
+  'yargs',
 ];
 
+const documentation = {
+  ruleName: 'require-no-unnecessary-dependency',
+  defaultValue: 'warn',
+  description: "This rule doesn't allow any unnecessary dependencies and peerDependencies to be passed in the package.json. Doesn't apply for devModule.",
+};
+
 module.exports = {
-  create: ({ ruleConfig, projectType, report }) => ({
+  create: ({
+    ruleConfig,
+    projectType,
+    report,
+  }) => ({
     dependencies: (dependencies) => {
       const messageString = 'require-no-unnecessary-dependencies';
       if (projectType !== 'devModule') {
         const currentProblems = Object.keys(dependencies).map((dependency) => {
           const dependencyVersion = dependencies[dependency];
-          if (devDependencySet.includes(dependency)) {
+          if (devDependencySet.includes(dependency) && !(ruleConfig.severity.allowList && ruleConfig.severity.allowList.includes(dependency))) {
             return `${dependency}@${dependencyVersion} does not satisfy requirement for ${messageString} rule.`;
           }
           return undefined;
@@ -89,7 +96,10 @@ module.exports = {
         if (currentProblems.length) {
           const lintMessage = `This project has unnecessary dependencies that violates the ${messageString} rule:\n  ${currentProblems.join('\n  ')}`;
           report({
-            lintId: messageString, severity: ruleConfig.severity.severityType, lintMessage, projectType,
+            lintId: messageString,
+            severity: ruleConfig.severity.severityType,
+            lintMessage,
+            projectType,
           });
         }
       }
@@ -99,7 +109,7 @@ module.exports = {
       if (projectType !== 'devModule') {
         const currentProblems = Object.keys(peerDependencies).map((dependency) => {
           const dependencyVersion = peerDependencies[dependency];
-          if (devDependencySet.includes(dependency)) {
+          if (devDependencySet.includes(dependency) && !(ruleConfig.severity.allowList && ruleConfig.severity.allowList.includes(dependency))) {
             return `${dependency}@${dependencyVersion} does not satisfy requirement for ${messageString} rule.`;
           }
           return undefined;
@@ -108,10 +118,14 @@ module.exports = {
         if (currentProblems.length) {
           const lintMessage = `This project has unnecessary peerDependencies that violates the ${messageString} rule:\n  ${currentProblems.join('\n  ')}`;
           report({
-            lintId: messageString, severity: ruleConfig.severity.severityType, lintMessage, projectType,
+            lintId: messageString,
+            severity: ruleConfig.severity.severityType,
+            lintMessage,
+            projectType,
           });
         }
       }
     },
   }),
+  documentation,
 };