+++ packages/package-json-lint/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Added
+  * Added new rule `require-no-unnecessary-dependency`.
+
+* Added
   * Added new rule `require-no-hard-coded-dependency-versions`.
 
 ## 1.1.0 - (August 31, 2021)