+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Added
+  * Added documentation for new rule `require-no-unnecessary-dependency`.
+
+* Added
   * Added documentation for `package-json-lint`.
 
 ## 2.4.0 - (September 2, 2021)